ajout livre ok api/books

+++ backend/controllers/book.js
@@ -1,9 +1,15 @@
 const Book = require ('../models/Books');
 
 exports.createBook = (req, res, next)=>{
-    delete req.body._id;
+    console.log('Requête reçue:', req.body);
+    const bookObject = JSON.parse (req.body.book);
+    
+    delete bookObject._id;
+    delete bookObject._userId; 
+
     const book = new Book({
-        ...req.body
+        ...bookObject,
+        imageUrl : `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     });
     book.save()
         .then(()=> res.status(201).json({message: 'livre enregistré'}))
@@ -73,14 +79,39 @@ exports.bestRating = (req, res, next) =>{
 }
 
 exports.modifyBook = (req,res,next)=>{
-    Book.updateOne({_id:req.params.id}, {...req.body,_id : req.params.id})
-        .then(()=> res.status(200).json ({message : 'livre modifié'}))
-        .catch(error => res.status(400).json ({error}));
+    const bookObject = req.file ? {
+        ...JSON.parse (req.body.book),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    } : { ...req.body};
+
+    delete bookObject._userId;
+    Book.findOne({_id: req.params.id})
+        .then((book)=>{
+            if (book.userId != req.auth.userId){
+                res.status(401).json ({message :'Non autorisé'});
+            } else {
+                Book.updateOne ({_id:req.params.id}, {...bookObject, _id: req.params.id})
+                    .then(()=> res.status(200).json ({message : "Objet modifié"}))
+                    .catch(error => res.status(401).json({error}));
+            }
+        })
+        .catch((error) => res.status(400).json ({error}));
 };
 
 exports.deleteBook = (req, res, next)=>{
-    Book.deleteOne ({_id: req.params.id})
-        .then(()=> res.status(200).json({message :'livre supprimé'}))
-        .catch(error => res.status(400).json ({error}));
+    Book.findOne({_id: req.params.id})
+        .then(book =>{
+            if (book.userId != req.auth.userId){
+                res.status(400).json ({message : 'Non autorisé'});
+            } else {
+                const filename = book.imageUrl.split ('/images') [1];
+                fs.unlink(path.join ('images',filename),()=>{
+                    Book.deleteOne({_id:req.params.id})
+                        .then(()=> res.status(200).json ({message : 'objet supprimé'}))
+                        .catch(error => res.status(401).json({error}))
+                });
+            }
+        })
+        .catch (error => { res.status(500).json ({error})});
 };
